Add files via upload
The present program instructs a Raspberry Pi Pico (original standard) to act as a slave over four wire SPI pins. The slave also takes readings from an external NTC thermistor through the ADC0 pin GP26, and turns three LEDs on and off from GPIO 0, GPIO 1, and GPIO 2. The slaves waits for a signal from the master. This can be a fraction of a second, or many seconds, it doesn't care. The slave is not in charge of pacing. The slave receives a byte from the master with the value, if any, of a LED to be toggled. This byte will have only one LED value at a time (4 = red, 2 = yellow, 1 = green), but conceivably any value could represent any combination of LEDs (e.g., 7 = red and yellow and green). It's just not written that way in the present version. Conversely, the slave sends a LED state byte to the master that can be in any combination of LED states, 0-7.

Core 1 is in charge of all processes having to do with the ADC and thermistor. The ADC of the Pico operates on a 12-bit value, so the data is sent to Core 0 using two bytes, which means the remaining four MSB will be 0000. The slave, then, sends a total of three bytes to the master: Byte 0 & Byte 1 are the ADC values. The master passively accepts these, without reply. Byte 2 is the accessory byte with the encoded LED states. 

The slave expects and receives only one byte from the master, telling it which LED to turn on or off, if any.